fix(release): satisfy stable core quality gates
